- (snmplib/asn1.c snmplib/snmp.c snmplib/snmp_api.c snmplib/snmpusm.c):   - Don't assume size_t is signed (Pekka Kytolaakso)
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@3775 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/snmplib/snmpusm.c
@@ -1406,7 +1406,7 @@ usm_check_and_update_timeliness(
 	int    *error)
 {
 	u_char	myID[USM_MAX_ID_LENGTH];
-	size_t	myIDLength = snmpv3_get_engineID(myID, USM_MAX_ID_LENGTH);
+	int	myIDLength = snmpv3_get_engineID(myID, USM_MAX_ID_LENGTH);
 	u_int	myBoots;
 	u_int	myTime;
